requirements.txt
streamlit
pandas
numpy
langchain-community
langchain-google-genai
python-dotenv